[CHORE] 오류 수정

+++ src/main/kotlin/com/official/memento/todo/controller/dto/PrioritizationResponse.kt
@@ -89,5 +89,5 @@ data class ToDoPrioritizedGetResponse(
         @Schema(description = "ToDo 유형")
     val toDoType: ToDoType,
         @Schema(description = "정렬 순서")
-    val orderNum: Int,
+    val orderNum: Double,
 )
